+++ src/Http/Controllers/SitemapController.php
@@ -10,21 +10,12 @@
 
 namespace Anubarak\Sitemap\Http\Controllers;
 
-use Craft;
-use craft\helpers\DateTimeHelper;
-use craft\helpers\FileHelper;
-use craft\helpers\UrlHelper;
-use craft\web\Controller;
-use Anubarak\Sitemap\helpers\PathHelper;
-use Anubarak\Sitemap\models\SitemapEntryModel;
-use Anubarak\Sitemap\records\SitemapCrawlerVisit;
-use Anubarak\Sitemap\Plugin;
-use DOMDocument;
-use Exception;
-use Jaybizzle\CrawlerDetect\CrawlerDetect;
-use secondred\formbuilder\elements\db\EntryQuery;
-use yii\web\HttpException;
-use yii\web\Response;
+use Anubarak\Sitemap\Sitemap;
+use Anubarak\Sitemap\Support\PathHelper;
+use CraftCms\Cms\Site\Sites;
+use CraftCms\Cms\Support\DateTimeHelper;
+use CraftCms\Cms\Support\File;
+use Symfony\Component\HttpKernel\Exception\HttpException;
 
 /**
  * Default Controller
@@ -35,6 +26,13 @@ use yii\web\Response;
 class SitemapController
 {
 
+    public function __construct(
+        private readonly Sites $sites,
+        private readonly Sitemap $sitemap,
+    )
+    {
+    }
+
     /**
      * Handle a request going to our plugin's index action URL,
      * e.g.: actions/sitemap/default
@@ -42,38 +40,36 @@ class SitemapController
      * @param string $suffix
      *
      * @return mixed
-     * @throws \yii\base\Exception
-     * @throws \craft\errors\SiteNotFoundException
      */
     public function index(string $suffix = '')
     {
         // grab the site-map
-        $site = Craft::$app->getSites()->getCurrentSite();
+        $site = $this->sites->getCurrentSite();
         $path = PathHelper::getSiteMapPath();
+
         // only default route
         $name = 'sitemap_' . $site->id  . $suffix . '.xml';
         if(file_exists($path . $name)){
-            $date = FileHelper::lastModifiedTime($path . $name);
+            $date = File::lastModified($path . $name);
             $date = DateTimeHelper::toDateTime($date);
             // older than a week? regenerate by force
             $now = new \DateTime();
             if($now->modify('-1 week') > $date){
                 // rebuild
-                Plugin::getInstance()->getSiteMap()->buildIndexFile($site);
+                $this->sitemap->buildIndexFile($site);
             }
 
             $xmlString = file_get_contents($path . $name);
         }else{
             if($suffix){
-                return $this->redirect('sitemap.xml');
+                return redirect()->route('sitemap-index');
             }else{
                 throw new HttpException(400, 'No sitemap found -> you need to generate it first via console command „php craft secondred-sitemap“');
             }
         }
 
-        $this->response->format = Response::FORMAT_RAW;
-        $this->response->getHeaders()->add('Content-Type', 'application/xml');
-
-        return $xmlString;
+        return response($xmlString, 200, [
+            'Content-Type' => 'application/xml'
+        ]);
     }
 }